requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
gspread>=5.12.4
google-auth>=2.29.0
